+++ router/httpdate.go
@@ -1,4 +1,4 @@
-package camo
+package router
 
 import (
 	"sync"